Use floats in js plot blog

+++ js-plot/algebra.cpp
@@ -1,9 +1,9 @@
 // An example equation
-double equation(double x) {
+float equation(float x) {
   return 2 * x * x * x - 4 * x * x + 6;
 }
 
 // Derivative of the above equation
-double derivative(double x) {
+float derivative(float x) {
   return 6 * x * x - 8 * x;
 }
